post request for allusers from getallpersonal page to webapi completed

+++ Appointments.Application/MediatR/Handlers/DeleteAllUserOperationHandler.cs
@@ -1,6 +1,8 @@
 ﻿using Appointments.Application.IRepositories;
 using Appointments.Application.MediatR.Requests.UserRequests;
 using Appointments.Application.MediatR.Responses.UserReponses;
+using Appointments.Domain.Enums;
+using Appointments.Shared;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -21,6 +23,14 @@ namespace Appointments.Application.MediatR.Handlers
 
 		public async Task<DeleteAllUserResponse> Handle(DeleteAllUserRequest request, CancellationToken cancellationToken)
 		{
+			var requesterUser = await userRepository.GetSingleAsync(x => x.Id == request.RequesterId);
+
+			if (requesterUser is null)
+				throw new Exception(Constants.USER_NOT_FOUND);
+
+			if (requesterUser.Role != UserRoleType.Admin)
+				throw new Exception(Constants.USER_NOT_PERMISSION);
+
 			await userRepository.DeleteAllAsync();
 
 			return new DeleteAllUserResponse();
